+ Updating FFMPEG

+++ lib/oexlib/inc/http_session.h
@@ -315,7 +315,7 @@ public:
 
 		// Do we have all the data?
 		oexLONG lContentLength = m_pbRxHeaders[ "content-length" ].ToLong();
-		if ( 0 < lContentLength && lContentLength > Rx().GetMaxRead() )
+		if ( 0 < lContentLength && lContentLength > (long)Rx().GetMaxRead() )
 			return 0;
 
 		// Set default headers
